Update inline documentation

+++ tests/functions/arkloneLogger.sh
@@ -1,12 +1,29 @@
 #!/bin/bash
+# arklone cloud sync utility
+# by ridgek
+# Released under GNU GPLv3 license, see LICENSE.md.
+
 source "/opt/arklone/config.sh"
 source "${ARKLONE[installDir]}/functions/arkloneLogger.sh"
 
+#####
+# RUN
+#####
 # Create test log
 arkloneLogger "/dev/shm/test.log"
 
+[ $? = 0 ] || exit $?
+
+########
+# TEST 1
+########
 # Log file exists
 [ -f "/dev/shm/test.log" ] || exit 73
 
-# Teardown
+echo "TEST 1 passed."
+
+##########
+# TEARDOWN
+##########
 rm "/dev/shm/test.log"
+
